Associate integerLiterals with String rather than UInt32.
I think I got the wrong end of the stick earlier. The Lexer isn't
supposed to be doing *any* parsing: it's not supposed to be extracting
the integer literal out of the character stream. But just associating a
part of the stream with a token type.

This works for now but I think I want to consider pulling the entire
file into memory and having the tokens with associated values have
references into that memory... Hm!