Use our new Assembly Lexer in the generator tests.
Instead of relying on String comparisons to check that the abstract
syntax tree is transformed into the correct assembly, tokenise the
assembly and verify that the correct tokens are produced.
This is much more accurate and means we don't have to worry about
indentations and other whitespace issues when writing the tests.