Improve the Generator with a helper class X86_64Builder.
Instead of writing raw assembly strings in the generator class, move
that work over to the X86_64Builder, which emits the string for each
type of assembly token.

Also stop using 'self.' where it's not necessary. The argument is lost
and it's time to start writing more idiomatic Swift code! I'll start
removing it from this project.